etl/requirements.txt
pandas
numpy
sqlalchemy
pytesseract
pdf2image
Pillow
psycopg2-binary
